Viz folders and files naming changes
