+++ src/App.js
@@ -4,7 +4,7 @@ import { Navigate, Route, BrowserRouter as Router, Routes } from "react-router-d
 import Home from './Components/Pages/Home';
 import Team from './Components/Pages/Team';
 // TODO: Change the below from index2 to index for Visualizer
-import Visualizer from './Components/Pages/Visualizer/index2';
+import Visualizer from './Components/Pages/Visualizer';
 import Footer from './Components/Shared/Footer';
 import Loader from './Components/Shared/Loader';
 import Navbar from './Components/Shared/Navbar';
